requirements.txt
streamlit>=1.45.0
lmstudio>=1.3.0
pyttsx3==2.90
SpeechRecognition>=3.10.0
Pillow>=10.0.0
pypiwin32; sys_platform == 'win32'
langchain
langchain-community
pypdf>=3.17.0
sentence-transformers>=2.2.2
torch>=2.0.0
faiss-cpu>=1.7.4
pyaudio
langchain-openai
opencv-python-headless